IR/lr02_crawler/crawler/requirements.txt
pymongo==4.6.1
pyyaml==6.0.1
requests==2.31.0
beautifulsoup4==4.12.2
fake-useragent==1.4.0